requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-docx>=1.1.0
markdown>=3.5
jinja2>=3.1.0
nltk>=3.8
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
openai>=1.0.0
pypdf>=4.0.0